requirements.txt
pymongo[srv]
streamlit-authenticator
openai-whisper
pydub
audio-recorder-streamlit